backend/requirements/base.txt
# Master branches
https://github.com/django/django/archive/main.tar.gz
https://github.com/encode/django-rest-framework/archive/master.tar.gz

whitenoise==5.2.0
djangorestframework-simplejwt==5.1.0